feat: integrate articles section properly into home page
